requirements.txt
absl-py==0.9.0
alembic==1.4.2
appdirs==1.4.4
asn1crypto==0.24.0
astunparse==1.6.3
atpublic==1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
boto3==1.18.59
botocore==1.21.59
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
cryptography==2.3
cycler==0.10.0
cymem==2.0.3
cytoolz==0.11.0
databricks-cli==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docker==4.2.0
docutils==0.15.2
dpath==2.0.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
grandalf==0.6
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
humanize==2.4.0
idna==2.6
importlib-metadata==1.6.0
inflect==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.15.1
jsonpath-ng==1.5.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.2
nanotime==0.5.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.3
nltk==3.6.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plac==1.1.3
ply==3.11
preshed==3.0.2
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pygtrie==2.3.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-pretrained-bert==0.6.1
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
querystring-parser==1.2.4
regex==2021.10.8
requests==2.23.0
requests-oauthlib==1.3.0
rouge==1.0.0rc2
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.5.0
sacremoses==0.0.43
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.91
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
smmap==3.0.4
spacy==2.2.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
thinc==7.4.0
tokenizers==0.7.0
toolz==0.11.1
torch==1.4.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
transformers==2.11.0
treelib==1.6.1
urllib3==1.25.9
voluptuous==0.11.7
wasabi==0.6.0
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.1.0
